cleanup resharding tests
